(refactor) Move test_gateway_in_flight_order.py
